feat: Add Collatz iterator implementation
Add a new iterator implementation called Collatz, which generates the Collatz sequence for a given number. The Collatz sequence is generated by repeatedly applying a set of rules until the number reaches 1. This iterator allows iterating over the sequence and counting the number of steps required to reach 1 for a given number.